backend/requirements.txt
Django>=3.2
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.7.2